refactor: place Next font manifest adapter
